javascript: Fix loading of image files
In Firefox (under Linux at least) it is not possible to get the full
path of a selected file from the '<input type="file">' element.

This fix uses a FileReader to read the content of the selected file as
binary and set the 'src' attribute of the '<img>' element in a similar
way.